+++ changelog.md
@@ -1,3 +1,9 @@
+**Loot Status**
+* Added loot status to the session frame.
+* This allows the ML to see how many have looted the boss before starting a session.
+
+
+
 ### v2.10.0
 ---
 * **Version Checker**