+++ changelog.md
@@ -2,8 +2,11 @@
 * Added loot status to the session frame.
 * This allows the ML to see how many have looted the boss before starting a session.
 
-* Added `notes` as an option to announce awards. 
+* Added `notes` as an option to announce awards.
 
+
+###### Bugfixes
+* *Fixed random issue with realm name (#385).*
 
 
 ### v2.10.0